more organization

+++ src/main/java/com/minecraftabnormals/buzzier_bees/client/render/layer/MoobloomButtercupLayer.java
@@ -26,16 +26,16 @@ public class MoobloomButtercupLayer<T extends MoobloomEntity> extends LayerRende
          BlockState blockstate = entitylivingbaseIn.getFlower().getDefaultState();
          int i = LivingRenderer.getPackedOverlay(entitylivingbaseIn, 0.0F);
          matrixStackIn.push();
-         matrixStackIn.translate((double)0.2F, (double)-0.35F, 0.5D);
+         matrixStackIn.translate(0.2F, -0.35F, 0.5D);
          matrixStackIn.rotate(Vector3f.YP.rotationDegrees(-48.0F));
          matrixStackIn.scale(-1.0F, -1.0F, 1.0F);
          matrixStackIn.translate(-0.5D, -0.5D, -0.5D);
          blockrendererdispatcher.renderBlock(blockstate, matrixStackIn, bufferIn, packedLightIn, i);
          matrixStackIn.pop();
          matrixStackIn.push();
-         matrixStackIn.translate((double)0.2F, (double)-0.35F, 0.5D);
+         matrixStackIn.translate(0.2F, -0.35F, 0.5D);
          matrixStackIn.rotate(Vector3f.YP.rotationDegrees(42.0F));
-         matrixStackIn.translate((double)0.1F, 0.0D, (double)-0.6F);
+         matrixStackIn.translate(0.1F, 0.0D, -0.6F);
          matrixStackIn.rotate(Vector3f.YP.rotationDegrees(-48.0F));
          matrixStackIn.scale(-1.0F, -1.0F, 1.0F);
          matrixStackIn.translate(-0.5D, -0.5D, -0.5D);
@@ -43,7 +43,7 @@ public class MoobloomButtercupLayer<T extends MoobloomEntity> extends LayerRende
          matrixStackIn.pop();
          matrixStackIn.push();
          this.getEntityModel().getHead().translateRotate(matrixStackIn);
-         matrixStackIn.translate(0.0D, (double)-0.7F, (double)-0.2F);
+         matrixStackIn.translate(0.0D, -0.7F, -0.2F);
          matrixStackIn.rotate(Vector3f.YP.rotationDegrees(-78.0F));
          matrixStackIn.scale(-1.0F, -1.0F, 1.0F);
          matrixStackIn.translate(-0.5D, -0.5D, -0.5D);
